requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
alembic
email-validator
